Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

Ok, fixed the commit message for the new change in gerrit.  The old change was abandonned, but this is doing the same thing but with the two fixes Steve suggested.  Also added IE6 and Linux Desktop.  Didn't see a direct need for Custom (too cumbersome to type in on a mobile device IMO), if it's needed then please specify.

Patch-set: 2