Browser: Remove duplicate actions from overflow menu when using quick controls
Removed New tab and Bookmarks buttons, and removed Forward when using extended
controls. Left Refresh in as that item alternates with Stop.
Removed New tab and New incognito tab from the nav (tabs) screen, as they have
buttons on the action bar.

Also anchored the menu lower down on the screen so it won't clip with the system
status bar.

Change-Id: Ic320c740ce513d8c687913df1956200f1570300a