Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

You are almost correct.  I was initially using chrome when I checked which I forgot will automatically redirect when https is not available (with an extension).  
While I'm not sure if it matters or not there are two others that support https at this time.  

https://www.nate.com/
https://szukaj.onet.pl/

Patch-set: 1