Update patch set 1
Patch Set 1: No score

I think one paradigm which doesn't neatly fit into the Android multi-tasking model is gaming, and with CM10 increasingly applying to tablets where memory hungry, graphically intense games are increasingly popular, it's worth considering that an increasing nunber of users will want to "clear the decks" before beginning a gaming session. Whether it's the job of individual applications to allow themselves to be shut down via their menus or whether third party solutions are sufficient is moot, but it's a point for consideration in this discussion.

Patch-set: 1