Update patch set 1
Patch Set 1: Code-Review-1

I looked into this and the files you changed are generated with a python script. If these strings are going to be changed the script must be also. I couldn't run the script and the whole thing seemed contradictory (a whole lot saying do not translate yet there are translations - this is why the warnings are there). Fixing this would probably require renaming all those xml files as well.

Patch-set: 1