Update patch set 1
Patch Set 1: Code-Review+1

XML file names are meaningless - we have seen that in the refactoring to the new cm_*.xml format.
I always said that if there are actual errors in AOSP translations, I would fix them. I think this change is valid, since it's pointless to have it set untranslatable, even when it exists in every other languages. Besides, the ordering of the strings is not changed and if that particular line would be changed by AOSP, it would be by removing the untranslatable setting. Then this won't cause merge conflicts

Patch-set: 1