BrowserSettings: update desktop UA string to match Chrome 33
Squashed commit of the following:

BrowserSettings: update desktop UA string to match Chrome 30

Now that we're using a Chromium-based WebView which closely matches
Chrome 30, we should report as a newer Chrome version when masquerading
as a desktop browser.

Change-Id: I301a8cfac9a83e66a6749872943e80ff5c4740a1

BrowserSettings: update desktop UA string to match Chrome 33

This matches the corresponding update to Chromium 33 for Android WebView.

Change-Id: Id2a8a832c2a637f7c5284d5006b6f3475bfba560