Update patch set 1
Patch Set 1:

Was really hoping that we could pick the stuff up and fix the issues, because there are a lot of good features there. Like most of the stuff in the QRD branches, there is a lot of good stuff but it's very unpolished.

Patch-set: 1
Reviewer: Steve Kondik <32@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0