Update patch set 1
Patch Set 1:

Steve, I think Scott nailed it: The useful-to-crap ratio is - IMHO - too low in QRD to just blindly merge it and clean it afterwards. In their UI code that I've seen so far, useful patches definitely aren't the majority, even setting aside the usual string problems. The browser code merge in question is a good example: annoying popups when leaving the browser, an overflow menu that's scrolling because it's so huge, an about dialog that prints the user agent (why does any user need this?), a download location selection that's crashing left and right (and a questionable idea anyway) etc. pp. All of this is why I think we should pick the _really_ useful bits instead of just dumping their stuff in.

Patch-set: 1