Update patch set 2
Patch Set 2:

Steven, actually the text zoom accessibility setting is broken (it is ignored because TEXT_AUTOSIZING). The NARROW_COLUMNS flag was ignored but the accessibility works as intended, so the patch is for that user that staill wants it.

Patch-set: 2
Reviewer: Jorge Ruesga <3134@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0