Update patch set 2
Patch Set 2:

Point is, though, it makes no sense to set a flag that we know does nothing -- you may as well just not set any flag at all.

Also, you've effectively introduced a preference whose only effect is "ignore the setting of another preference".  This makes no sense at all; why, as a user, should I select this new preference instead of unsetting the other preference directly?

As far as I'm concerned, what really should happen is one of two things:

(1) the accessibility feature is reimplemented to be compatible with text autosizing (it's worth noting that Chrome seems to have a superficially similar feature); or
(2) the Preferences UI is changed to make it clear that the two features are incompatible (selecting one disables and grays out the other, for example).

Patch-set: 2