Update patch set 3
Patch Set 3:

The blue menu looks ugly IMHO, and I haven't seen that in any of the Google apps so far. IMHO the action bar should be blue (if anything) and the menu should be either light or dark background. Also the folder icon doesn't really fit the tab menu IMHO; it would be better to just recolor the existing icon to white.

About mdpi, I think some of the tablets we support are mdpi.

Patch-set: 3