Update patch set 13
Patch Set 13:

Can anyone please make a screenshot with the launcher icon in PS12? I can't look at it from the phone :(

In general, I dislike the 'put something on a circular background' approach. _None_ of Google's apps does that, which is why the browser icon in e.g. PS6 looks a little out of place.

Patch-set: 13