Update patch set 13
Patch Set 13:

The black search icon is the one used with the search suggestion but i think i can switch to the one u find in the pie controls. I can remove that trasparency stuff on the tabs too, same for accent color. I'll also give a look at the bookmark theme.

Patch-set: 13