Update patch set 19
Patch Set 19:

More than 1000 users (1121 now) have voted and the most favourite icon is the world one (56%), followed by the current one (37%) and the compass one got 7% of votes. It's OK for you if I submit the world icon? And I'd like to add a fab to open a new tab, should I place it on the opened-tabs activity or over the webpage with auto-hiding effect (like g+)?

Patch-set: 19