Update patch set 1
Patch Set 1:

+1000 OMG!!!!

it does make sense in a way:
- For a lot of people, IE == Internet so it it really clear
- It is a fun easter egg (just not so well hidden)

Can't this be integrated in the browser settings menu with a switch of some kind? That would be really cool.
Or turn it on by default and watch the world burn ;)

Patch-set: 1
Reviewer: roel deckers <7097@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0