Update patch set 3
Patch Set 3:

It did not, that is a hidden class that can only be accessed if you build against CM directly. Since we are compiling this as a third party app, you can't access it without reflection. Calling out to the getprop binary will always work on our devices though, this is how most third party apps read these properties as far as I know.

Patch-set: 3