Update patch set 1
Patch Set 1:

This looks like an important check that was missing, but it seems like the stack trace for the crash was on the URI encode call, not on the replace. So, could this still happen? It says it tried to invoke length() on a null String. Which do we think was null? The query parameter? I'm not sure. Trying to build it myself to test.

Patch-set: 1