Update patch set 1
Patch Set 1:

The reason why URI encode call was failing is because not all search engine URIs have a requirement for replacing PARAMETER_PARTNER_CODE like for Bing: http://api.bing.com/osjson.aspx?FORM=CYANDF&amp;PC={partnerCode}&amp;query={searchTerms}&amp;language={language}, but Baidu's is: http://www.baidu.com/s?wd={searchTerms. From lines 102 to 110, partnerCodeKey is initialized as null, but since the search uri does not contain either Yahoo or Bing for other search engines, the partnerCodeKey remains null. Then in lines 112 to 118, it only attempts to populate mPartnerCode if partnerCodeKey is not null, so that remains null as well since it was never initialized. When mPartnerCode is finally passed to encodeUri, it is attempting a .length() on null.

Patch-set: 1