Update patch set 1
Patch Set 1: Code-Review+2

ahh yes that makes sense, thank you! I think I was concentrating on line 217 (before patch) and was trying to think of why the query was null but the trace was actually for 225. My mistake, this makes perfect sense, +2 and sorry for writing this bug in the first place :(

Patch-set: 1
Label: Code-Review=+2