Browser: Resurface incognito  button
The current method of entering incognito mode is too discreet, and
there is no way of knowing how to enter it (long press new tab FAB).

Resurface the incognito button in the tab navigation screen.

Furthermore, on tablets, there is no way to enter incognito mode
AFAIK (or it is ridiculously difficult to find).

On tablets, add an entry to the overflow menu for creating new
incognito tabs.

Change-Id: I1c81d2addd16c11480d978aebf07336307ec694f