feijao: update favicon and thumbnail empty items
number of items in the bookmark_preloads in bookmarks_icons.xml should be
same as number of items of bookmarks in strings.xml otherwise exception
occurs and it failed to load the bookmarks.
This patch defines the empty item for bookmarks icons and favicon
since nothing is provided/requested from customer.

Issue-id: FEIJ-48
Change-Id: I8359c02aad5be4e7eccac11e7bab0388cd28401a